LPS-200228 Rename to commands for consistency with other modules
