+++ modules/apps/layout/layout-page-template-admin-web/src/main/resources/META-INF/resources/js/ActionsComponentPropsTransformer.js
@@ -5,7 +5,7 @@
 
 import {openSelectionModal} from 'frontend-js-web';
 
-import openDeletePageTemplateModal from './modal/openDeletePageTemplateModal';
+import openDeletePageTemplateModal from './commands/openDeletePageTemplateModal';
 
 const ACTIONS = {
 	deleteCollections({
