+++ modules/apps/layout/layout-page-template-admin-web/src/main/resources/META-INF/resources/js/commands/openDeletePageTemplateModal.js
