+++ modules/apps/layout/layout-page-template-admin-web/src/main/resources/META-INF/resources/js/propsTransformers/MasterLayoutManagementToolbarPropsTransformer.js
@@ -5,7 +5,7 @@
 
 import {openSimpleInputModal} from 'frontend-js-web';
 
-import openDeletePageTemplateModal from '../modal/openDeletePageTemplateModal';
+import openDeletePageTemplateModal from '../commands/openDeletePageTemplateModal';
 
 export default function propsTransformer({portletNamespace, ...otherProps}) {
 	const deleteSelectedMasterLayouts = () => {
